backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
pydantic
google-search-results
pypdf